Time: 33 ms (54.65%) | Memory: 105 MB (52.91%) - LeetSync

+++ 3958-minimum-removals-to-balance-array/minimum-removals-to-balance-array.cpp
@@ -7,7 +7,7 @@ public:
         int i=0,j=0;
         int ans=0;
         while(j<n){
-            if((long long)nums[j]<=(long long)k*nums[i]){
+            if(nums[j]<=(long long)k*nums[i]){
                 ans=max(ans,j-i+1);
                 j++;
             }
